rename: Profit -> ProfitCalculator 변경
Profit은 도메인 보다 utils 에 가깝다고 생각하여, 클래스 이름 변경 후
패키지 변경을 진행하였음
